+++ packages/common/web/src/consts.js
@@ -3,34 +3,6 @@ export const HEADER_BLOCK = Object.freeze({
   TWO: 'header-two',
   THREE: 'header-three',
 });
-
-export const BLOCK_TYPES = Object.freeze([
-  'header-one',
-  'header-two',
-  'header-three',
-  'header-four',
-  'header-five',
-  'header-six',
-  'unordered-list-item',
-  'ordered-list-item',
-  'blockquote',
-  'atomic',
-  'code-block',
-  'unstyled',
-]);
-
-export const TITLE_FONT_STYLE = {
-  H1: 'header-one',
-  H2: 'header-two',
-  H3: 'header-three',
-  H4: 'header-four',
-  H5: 'header-five',
-  H6: 'header-six',
-  P: 'unstyled',
-};
-
-export const DEFAULT_FONTS_OPTIONS = ['P', 'H2', 'H3'];
-
 export const BUTTON_TYPE = 'wix-draft-plugin-button';
 export const CODE_BLOCK_TYPE = 'code-block';
 export const DIVIDER_TYPE = 'wix-draft-plugin-divider';