+++ packages/editor-common/web/src/consts.js
@@ -1,3 +1,30 @@
+export const BLOCK_TYPES = Object.freeze([
+  'header-one',
+  'header-two',
+  'header-three',
+  'header-four',
+  'header-five',
+  'header-six',
+  'unordered-list-item',
+  'ordered-list-item',
+  'blockquote',
+  'atomic',
+  'code-block',
+  'unstyled',
+]);
+
+export const TITLE_FONT_STYLE = {
+  H1: 'header-one',
+  H2: 'header-two',
+  H3: 'header-three',
+  H4: 'header-four',
+  H5: 'header-five',
+  H6: 'header-six',
+  P: 'unstyled',
+};
+
+export const DEFAULT_FONTS_DROPDOWN_OPTIONS = ['P', 'H2', 'H3', 'H4', 'H5', 'H6'];
+
 export const COMMANDS = Object.freeze({
   TITLE: 'header-two',
   SUBTITLE: 'header-three',